+++ src/BridgeMaker.js
@@ -1,4 +1,4 @@
-const { DEFAULT } = require("./constant.js");
+const { DEFAULT } = require("../constant.js");
 const BridgeMaker = {
   /**
    * @param {number} size 다리의 길이
