+++ src/BridgeRandomNumberGenerator.js
